SP-359 feat: 스터디 지원자 API service
